+++ src/Xilium.Crdtp.Client/tests/__RefactorMe_CrdtpSessionTests.cs
@@ -49,7 +49,7 @@ namespace Xilium.Crdtp.Client.Tests
             // TODO: Normally this code also should observe what SendAsync is not get called, nor message is serialized.
 
             var client = new CrdtpClient((handler) => new FakeConnection(handler));
-            var session = new CrdtpSession("");
+            var session = new CrdtpSession(client, "");
             client.Attach(session);
             var sessionApi = new TestSessionApi(session);
             await client.OpenAsync();
@@ -72,7 +72,7 @@ namespace Xilium.Crdtp.Client.Tests
             var client = new CrdtpClient((handler) => (connection = new FakeConnection(handler)));
             Assert.NotNull(connection);
 
-            var session = new CrdtpSession("");
+            var session = new CrdtpSession(client, "");
             client.Attach(session);
             var sessionApi = new TestSessionApi(session);
             await client.OpenAsync();
@@ -87,9 +87,7 @@ namespace Xilium.Crdtp.Client.Tests
             var actualException = await Assert.ThrowsAnyAsync<OperationCanceledException>(() => doCallTask);
 
             // Ensure what there is no pending requests remains.
-#pragma warning disable CS0612 // Type or member is obsolete
-            Assert.Equal(0, sessionApi.GetCrdtpSession().GetNumberOfPendingRequestsForTesting());
-#pragma warning restore CS0612 // Type or member is obsolete
+            Assert.Equal(0, sessionApi.GetCrdtpSession().GetNumberOfPendingRequests());
 
             await client.CloseAsync();
         }